# backend-swordhealth/src/config/swagger.yaml
swagger: "2.0"
info:
  description: "SwordHealth Challenge"
  version: "1.0.0"
  title: "swordhealth"
basePath: "/api/v1"
consumes:
  - "application/json"
produces:
  - "application/json"
...
definitions:
  UpdateTask:
    type: "object"
    required:
      - "id"
      - "user_id"
      - "name"
      - "summary"
      - "perfomed_task"
    properties:
      id:
        type: "number"
      user_id:
        type: "number"
      name:
        type: "string"
      summary:
        type: "string"
      perfomed_task:
        type: "string"
  Task:
    type: "object"
    required:
      - "id"
      - "user_id"
      - "name"
      - "summary"
      - "perfomed_task"
      - "created_at"
      - "updated_at"
    properties:
      id:
        type: "number"
      user_id:
        type: "number"
      name:
        type: "string"
      summary:
        type: "string"
      perfomed_task:
        type: "string"
      created_at:
        type: "string"
      updated_at:
        type: "string"
  TasksGetResponse:
    type: "object"
    required:
      - "tasks"
    properties:
      tasks:
        type: "array"
        items:
          $ref: "#/definitions/Task"
  CreateTask:
    type: "object"
    required:
      - "name"
      - "summary"
    properties:
      name:
        type: "string"
      summary:
        type: "string"
...
  Validation:
    type: "object"  
    properties: 
      message:
        type: "string"
      path:
        type: "array"
        items:
          type: "string"
      type:
        type: "string"
      context:
        type: "object"
        properties:
          limit:
            type: "number"
          value: 
            type: "string"
          label:
            type: "string"
          key:
            type: "string"
  BadRequest:
    type: "object"
    properties:
      details:
        type: "array"
        items:
          $ref: "#/definitions/Validation"
  DefaultResponse:
    type: "object"
    properties:
      message:
        type: "string"
paths:
...
  /tasks:
    post:
      tags:
        - 'tasks'
      x-swagger-router-controller: app
      summary: create one task
      operationId: createTask
      security:
        - Bearer: []
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "task object"
        required: true
        schema:
          $ref: "#/definitions/CreateTask"
      responses:
        201:
          description: task has been created
          schema:
            $ref: "#/definitions/DefaultResponse"
        400:
          description: BadRequest
          schema:
            $ref: "#/definitions/BadRequest"
        500:
          description: internalError - UUID to search for logs
          schema:
            $ref: "#/definitions/DefaultResponse"
    get:
      tags:
        - 'tasks'
      x-swagger-router-controller: app
      summary: get many tasks
      operationId: getTasks
      security:
        - Bearer: []
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "limit"
        in: "query"
        description: "limit data"
        type: "string"
      - name: "offset"
        in: "query"
        description: "page"
        type: "string"
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/TasksGetResponse"
        400:
          description: BadRequest
          schema:
            $ref: "#/definitions/BadRequest"
        500:
          description: internalError - UUID to search for logs
          schema:
            $ref: "#/definitions/DefaultResponse"
    put:
      tags:
        - 'tasks'
      x-swagger-router-controller: app
      summary: update one task
      operationId: updateTask
      security:
        - Bearer: []
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "task object"
        required: true
        schema:
          $ref: "#/definitions/UpdateTask"
      responses:
        204:
          description: task has been updated
          schema:
            $ref: "#/definitions/DefaultResponse"
        400:
          description: BadRequest
          schema:
            $ref: "#/definitions/BadRequest"
        500:
          description: internalError - UUID to search for logs
          schema:
            $ref: "#/definitions/DefaultResponse"
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: Bearer token